+++ src/AppserverIo/Appserver/PersistenceContainer/PersistenceManager.php
@@ -134,6 +134,7 @@ class PersistenceManager extends AbstractManager implements PersistenceContextIn
                 // append the application specific properties
                 $properties->add(SystemPropertyKeys::WEBAPP, $webappPath = $application->getWebappPath());
                 $properties->add(SystemPropertyKeys::WEBAPP_NAME, basename($webappPath));
+                $properties->add(SystemPropertyKeys::WEBAPP_DATA, $application->getDataDir());
                 $properties->add(SystemPropertyKeys::WEBAPP_CACHE, $application->getCacheDir());
                 $properties->add(SystemPropertyKeys::WEBAPP_SESSION, $application->getSessionDir());